+++ R/emmGrid-methods.R
@@ -49,13 +49,8 @@ str.emmGrid <- function(object, ...) {
         if (is.null(x)) cat("(predicted by other variables)")
         else cat(paste(format(x, digits = 5, justify = "none"), collapse=", "))
     }
-    showtran = function(tran, label) { # internal convenience fcn
-        if (is.list(tran)) 
-            tran = ifelse(is.null(tran$name), "custom", tran$name)
-        if (!is.null(mul <- object@misc$tran.mult))
-            tran = paste0(mul, "*", tran)
-        cat(paste(label, dQuote(tran), "\n"))
-        
+    showtran = function(misc, label) { # internal convenience fcn
+        cat(paste(label, dQuote(.fmt.tran(misc)), "\n"))
     }
     levs = object@levels
     cat(paste("'", class(object)[1], "' object with variables:\n", sep=""))
@@ -87,9 +82,9 @@ str.emmGrid <- function(object, ...) {
         cat("\n")
     }
     if(!is.null(tran <- object@misc$tran)) {
-        showtran(tran, "Transformation:")
+        showtran(object@misc, "Transformation:")
         if (!is.null(tran2 <- object@misc$tran2))
-            showtran(tran2, "Additional response transformation:")
+            showtran(list(tran = tran2), "Additional response transformation:")
     }
 }
 
@@ -175,6 +170,10 @@ vcov.emmGrid = function(object, ...) {
 #' response transformation \samp{2*sqrt(y)} (or \samp{sqrt(y) + sqrt(y + 1)},
 #' for that matter), we should have \code{tran = "sqrt"} and \code{tran.mult =
 #' 2}. If absent, a multiple of 1 is assumed.}
+#' 
+#' \item{\code{tran.offset}}{Additive constant before a transformation is applied.
+#' For example, a response transformation of \code{log(y + pi)} has
+#' \code{tran.offset  = pi}. If no value is present, an offset of 0 is assumed.}
 #' 
 #' \item{\code{estName}}{(\code{character}) is the column label used for
 #' displaying predictions or EMMs.}
@@ -288,7 +287,7 @@ update.emmGrid = function(object, ..., silent = FALSE) {
     valid.misc = c("adjust","alpha","avgd.over","by.vars","delta","df",
                    "initMesg","estName","estType","famSize","infer","inv.lbl",
                    "level","methDesc","nesting","null","predict.type","pri.vars"
-                   ,"side","sigma","tran","tran.mult","tran2","type","is.new.rg")
+                   ,"side","sigma","tran","tran.mult","tran.offset","tran2","type","is.new.rg")
     valid.slots = slotNames(object)
     valid.choices = union(valid.misc, valid.slots)
     misc = object@misc
@@ -689,10 +688,10 @@ regrid = function(object, transform = c("response", "mu", "unlink", "log", "none
         }
         if((transform %in% c("mu", "unlink")) && !is.null(object@misc$tran2)) {
             object@misc$tran = object@misc$tran2
-            object@misc$tran2 = object@misc$tran.mult = object@misc$inv.lbl = NULL
+            object@misc$tran2 = object@misc$tran.mult = object@misc$tran.offset = object@misc$inv.lbl = NULL
         }
         else
-            object@misc$tran = object@misc$tran.mult = object@misc$inv.lbl = NULL
+            object@misc$tran = object@misc$tran.mult = object@misc$tran.offset = object@misc$inv.lbl = NULL
         sigma = object@misc$sigma = NULL
     }
     if (transform == "log") { # from prev block, we now have stuff on response scale